Attempt to simplify changes ans reduce diff
