+++ kinto/core/errors.py
@@ -1,10 +1,11 @@
 import six
+import colander
 from pyramid import httpexceptions
 from enum import Enum
 
 from kinto.core.logs import logger
 from kinto.core.utils import json, reapply_cors, encode_header
-from kinto.core.schema import ErrorSchema
+from kinto.core.schema import Any
 
 
 class ERRORS(Enum):
@@ -71,6 +72,17 @@ class ERRORS(Enum):
     UNDEFINED = 999
     BACKEND = 201
     SERVICE_DEPRECATED = 202
+
+
+class ErrorSchema(colander.MappingSchema):
+    """Payload schema for Kinto errors."""
+
+    code = colander.SchemaNode(colander.Integer())
+    errno = colander.SchemaNode(colander.Integer())
+    error = colander.SchemaNode(colander.String())
+    message = colander.SchemaNode(colander.String(), missing=colander.drop)
+    info = colander.SchemaNode(colander.String(), missing=colander.drop)
+    details = colander.SchemaNode(Any(), missing=colander.drop)
 
 
 def http_error(httpexception, errno=None,
